docker-compose/old/requirements.txt
redis
psycopg2-binary
flask